api/serializers.py
from nuru.models import User, Categories, Men, Women, Kids
from rest_framework import serializers
class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("first_name", "last_name", "email", "password")